Consolidated repeated instances of MessageType.
